+++ components/style/gecko/wrapper.rs
@@ -19,7 +19,7 @@ use app_units::Au;
 use applicable_declarations::ApplicableDeclarationBlock;
 use atomic_refcell::{AtomicRefCell, AtomicRefMut};
 use context::{QuirksMode, SharedStyleContext, UpdateAnimationsTasks};
-use data::ElementData;
+use data::{ElementData, RestyleData};
 use dom::{self, DescendantsBit, LayoutIterator, NodeInfo, TElement, TNode, UnsafeNode};
 use dom::{OpaqueNode, PresentationalHintsSynthesizer};
 use element_state::{ElementState, DocumentState, NS_DOCUMENT_STATE_WINDOW_INACTIVE};
@@ -63,7 +63,9 @@ use gecko_bindings::structs::ELEMENT_HAS_SNAPSHOT;
 use gecko_bindings::structs::EffectCompositor_CascadeLevel as CascadeLevel;
 use gecko_bindings::structs::NODE_IS_IN_NATIVE_ANONYMOUS_SUBTREE;
 use gecko_bindings::structs::NODE_IS_NATIVE_ANONYMOUS;
+use gecko_bindings::structs::nsChangeHint;
 use gecko_bindings::structs::nsIDocument_DocumentTheme as DocumentTheme;
+use gecko_bindings::structs::nsRestyleHint;
 use gecko_bindings::sugar::ownership::{HasArcFFI, HasSimpleFFI};
 use logical_geometry::WritingMode;
 use media_queries::Device;
@@ -669,6 +671,64 @@ impl<'le> GeckoElement<'le> {
     /// Owner document quirks mode getter.
     pub fn owner_document_quirks_mode(&self) -> QuirksMode {
         self.as_node().owner_doc().mCompatMode.into()
+    }
+
+    /// Only safe to call on the main thread, with exclusive access to the element and
+    /// its ancestors.
+    /// This function is also called after display property changed for SMIL animation.
+    ///
+    /// Also this function schedules style flush.
+    unsafe fn maybe_restyle<'a>(&self,
+                                data: &'a mut ElementData,
+                                animation_only: bool) -> Option<&'a mut RestyleData> {
+        use dom::{AnimationOnlyDirtyDescendants, DirtyDescendants};
+
+        // Don't generate a useless RestyleData if the element hasn't been styled.
+        if !data.has_styles() {
+            return None;
+        }
+
+        // Propagate the bit up the chain.
+        if let Some(p) = self.traversal_parent() {
+            if animation_only {
+                p.note_descendants::<AnimationOnlyDirtyDescendants>();
+            } else {
+                p.note_descendants::<DirtyDescendants>();
+            }
+        };
+
+        bindings::Gecko_SetOwnerDocumentNeedsStyleFlush(self.0);
+
+        // Ensure and return the RestyleData.
+        Some(&mut data.restyle)
+    }
+
+    /// Set restyle and change hints to the element data.
+    pub fn note_explicit_hints(&self,
+                               restyle_hint: nsRestyleHint,
+                               change_hint: nsChangeHint) {
+        use gecko::restyle_damage::GeckoRestyleDamage;
+        use invalidation::element::restyle_hints::RestyleHint;
+
+        let damage = GeckoRestyleDamage::new(change_hint);
+        debug!("note_explicit_hints: {:?}, restyle_hint={:?}, change_hint={:?}",
+               self, restyle_hint, change_hint);
+
+        let restyle_hint: RestyleHint = restyle_hint.into();
+        debug_assert!(!(restyle_hint.has_animation_hint() &&
+                        restyle_hint.has_non_animation_hint()),
+                      "Animation restyle hints should not appear with non-animation restyle hints");
+
+        let mut maybe_data = self.mutate_data();
+        let maybe_restyle_data = maybe_data.as_mut().and_then(|d| unsafe {
+            self.maybe_restyle(d, restyle_hint.has_animation_hint())
+        });
+        if let Some(restyle_data) = maybe_restyle_data {
+            restyle_data.hint.insert(restyle_hint.into());
+            restyle_data.damage |= damage;
+        } else {
+            debug!("(Element not styled, discarding hints)");
+        }
     }
 }
